+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Introduce simpler request bodies for endpoints `POST /change-password`, `POST /token` and `POST /register`. Old
+  behavior can be disabled with the feature flag `featureFlag.280_OldUniqueUserIdentifierDisabled`. Closes #280.
+### Deprecated
+- Original request body scheme of endpoints `POST /change-password`, `POST /token` and `POST /register` is deprecated,
+  will be removed in version 0.2.0, part of #280.
 
 ## 0.1.5 - 2024-04-12
 ### Changed